Split website controller into listings and info
This is mostly a lot of renaming.

Note that the main partials from the listings were previously namespaced
under /events

Trash the website controller spec: it's almost useless and was done
before I had any clue what testing was.
