+++ app/sweepers/event_sweeper.rb
@@ -15,6 +15,6 @@ class EventSweeper < ActionController::Caching::Sweeper
 
   def expire_cache
     @controller ||= ActionController::Base.new
-    expire_fragment('website#index')
+    expire_fragment('listings#index')
   end
 end
