--- app/views/events/_class_listings.html.erb
@@ -1,5 +1,0 @@
-<ul class="daylist unstyled">
-<% DAYNAMES.each do |day| %>
-    <%= render :partial => "events/class_listing", :locals => {:day => day} %>
-<% end %>
-</ul>
