+++ app/views/info/about.html.haml
