+++ app/views/info/privacy.html
