+++ app/views/layouts/_google_analytics.html.erb
