+++ app/views/layouts/name_clash.html.haml
@@ -1,5 +1,5 @@
 - content_for :analytics do
-  = render "website/google_analytics"
+  = render "layouts/google_analytics"
 
 - content_for :stylesheet do
   = stylesheet_link_tag 'application', media: "screen"
