+++ app/views/listings/_advert.html.haml
