+++ app/views/listings/_class_listing.html.erb
