+++ app/views/listings/_class_listings.html.erb
@@ -1,0 +1,5 @@
+<ul class="daylist unstyled">
+<% DAYNAMES.each do |day| %>
+  <%= render "listings/class_listing", day: day %>
+<% end %>
+</ul>
