+++ app/views/listings/_navigation.html.erb
@@ -24,10 +24,10 @@
 </div>
 
 <div id="soldn_on_the_web">
-  <%= render 'website/advert', advert: ad %>
+  <%= render 'advert', advert: ad %>
 
   <div id="social_links">
-    <%= render 'website/facebook_buttons', text: "Like Swing Out London on Facebook", url: "https://www.swingoutlondon.co.uk" %>
+    <%= render 'facebook_buttons', text: "Like Swing Out London on Facebook", url: "https://www.swingoutlondon.co.uk" %>
   </div>
 </div>
