+++ app/views/listings/_social_listing.html.erb
@@ -7,7 +7,7 @@
     <% end %>
 
     <%if index==1%>
-       <%= render partial: "website/donate_message"%>
+       <%= render "listings/donate_message"%>
     <%end%>
     </ul>
 </li>
