+++ app/views/shared/_header_navigation.html.erb
@@ -1,1 +1,1 @@
-<%= render partial: 'website/navigation', locals: { nav_id: 'main_nav', nav_class: "navigation primary-nav", links: main_navigation_links } %>
+<%= render partial: 'layouts/website/navigation', locals: { nav_id: 'main_nav', nav_class: "navigation primary-nav", links: main_navigation_links } %>
